backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
dj-rest-auth==2.1.7
Django==3.2.10
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.0
djangorestframework-simplejwt==5.0.0
importlib-metadata==4.8.2
Markdown==3.3.6
Pillow==8.4.0
PyJWT==2.3.0
pymongo==4.0.1
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
wincertstore==0.2
zipp==3.6.0
